minor changes discussed in meeting:  changes related to nomenclature and moving of connection related classes to a new module named connection.py
